A couple of changes to manual config stuff to incorporate existence of api docs, lots of changes to makedocs, and adding makedocs to travis config.  crossing my fingers...

+++ makedocs.py
@@ -4,6 +4,7 @@ import argparse
 import subprocess
 import os.path
 import shutil
+import sys
 
 DEFAULT_TAG = '0.1'
 
@@ -35,7 +36,7 @@ def main(args):
     SPHINX_DIR=os.path.join(os.path.expanduser('~'),'__api-doc')
 
     #where should the temporary clone of the shakemap gh-pages repo live?
-    CLONE_DIR=os.path.join(os.path.expanduser('~'),'__shake-doc')
+    CLONE_DIR=os.path.join(os.path.expanduser('~'),'__shake-doc','html')
 
     #where is the repository where this script is running?
     REPO_DIR = os.path.dirname(os.path.abspath(__file__)) #where is this script?
@@ -58,24 +59,43 @@ def main(args):
 
     try:
         #clone the repository
+        sys.stderr.write('Cloning shakemap gh-pages branch...\n')
+        if os.path.isdir(CLONE_DIR):
+            shutil.rmtree(CLONE_DIR)
         clonecmd = 'git clone -b gh-pages https://github.com/usgs/shakemap.git %s' % CLONE_DIR
         res,stdout,stderr = getCommandOutput(clonecmd)
+        if not res:
+            raise Exception('Could not clone gh-pages branch.')
 
         #remove the current apidocs folder from that repo
-        apidocfolder = os.path.join(CLONE_DIR,'html','apidoc')
+        apidocfolder = os.path.join(CLONE_DIR,'apidoc') #this should probably not have 'html' in the path
         if os.path.isdir(apidocfolder):
             shutil.rmtree(apidocfolder)
 
         #run the make command to build the shakemap manual
+        sys.stderr.write('Building shakemap manual (HTML)...\n')
         makefile = os.path.join(REPO_DIR,'doc','Makefile')
         os.chdir(os.path.join(REPO_DIR,'doc'))
         manualcmd = '%s html' % make_cmd
         res,stdout,stderr = getCommandOutput(manualcmd)
+        if not res:
+            raise Exception('Could not build the ShakeMap manual.')
         
         #run the sphinx api doc command
+        sys.stderr.write('Building shakemap API documentation (REST)...\n')
         package_dir = os.path.join(REPO_DIR,'shakemap')
         sphinx_cmd = 'sphinx-apidoc -o %s -f -l -F -H %s -A "%s" -V %s %s' % (SPHINX_DIR,PACKAGE,AUTHORS,verstr,package_dir)
         res,stdout,stderr = getCommandOutput(sphinx_cmd)
+        if not res:
+            raise Exception('Could not build ShakeMap API documentation.')
+
+        #run the make command to build the shakemap manual (pdf version)
+        sys.stderr.write('Building shakemap manual (PDF)...\n')
+        os.chdir(os.path.join(REPO_DIR,'doc'))
+        manualcmd = '%s latexpdf' % make_cmd
+        res,stdout,stderr = getCommandOutput(manualcmd)
+        if not res:
+            raise Exception('Could not build the PDF version of the ShakeMap manual.')
 
         #this has created a conf.py file and a Makefile.  We need to "edit" the conf.py file to include the ReadTheDocs theme.
         fname = os.path.join(SPHINX_DIR,'conf.py')
@@ -87,8 +107,11 @@ def main(args):
         f.close()
 
         #Go to the sphinx directory and build the html
+        sys.stderr.write('Building shakemap API documentation (HTML)...\n')
         os.chdir(SPHINX_DIR)
         res,stdout,stderr = getCommandOutput('%s html' % make_cmd)
+        if not res:
+            raise Exception('Could not build HTML for API documentation.')
         
         #copy the generated content to the gh-pages branch we created earlier
         htmldir = os.path.join(SPHINX_DIR,'_build','html')
@@ -98,7 +121,8 @@ def main(args):
         
         if args.post:
             #cd to directory above where html content was pushed
-            os.chdir(CLONE_DIR)
+            htmldir = os.path.join(CLONE_DIR) #this should probably not have 'html' in it
+            os.chdir(htmldir)
             res,stdout,stderr = getCommandOutput('touch .nojekyll')
             res1,stdout,stderr = getCommandOutput('git add --all')
             res2,stdout,stderr = getCommandOutput('git commit -am"Pushing version %s to GitHub pages"' % verstr)
@@ -106,12 +130,13 @@ def main(args):
             if res1+res2+res3 < 3:
                 print('Something bad happened when attempting to add, commit, or push gh-pages content to GitHub. Exiting.')
                 sys.exit(1)
+            print('You can inspect the ShakeMap manual and API docs by looking here: http://usgs.github.io/shakemap/index.html')
         else:
             if not args.clean:
-                indexpage = os.path.join(CLONE_DIR,'html','index.html')
+                indexpage = os.path.join(CLONE_DIR,'index.html') #this should not have 'html' in it
                 print('You can inspect the ShakeMap manual and API docs by looking here: %s' % indexpage)
-    except:
-        pass
+    except Exception as error:
+        print('Error detected: %s' % str(error))
     finally:
         if args.clean:
             print('Cleaning up %s and %s created directories...' % (CLONE_DIR,SPHINX_DIR))
